config: set `tag_mappings` to an empty list as default (Bug 1964441) (#46)
We removed the `tag_mappings` entries in the production config file,
however the Pydantic schema requires them to be set. Set a default
value for the `tag_mappings` field, to an empty list.

Pydantic is different from dataclasses, in that you can set a
mutable type to a default value and it will correctly perform
a deep copy so each instance of the class has a new default.