test: use dedicated branch name to test branch mapping
The test configuration is uses both for unit and integration tests. So
far, we have been using the example of an ESR to test dynamic
replacements in url, branch and tags, which has been working as expected
across all tests.

However, with the recent changes separating each ref into their own
push. The configuration resulting in pushes to both the \`default\` and
\`mozilla-esr*\` branches creates an issue, as the second push does not
contain any new commit, and gets rejected by Mercurial.

This is not a pattern we currently want to support, so the simplest
approach is to split the parameters so we have dedicated branch-mapping
config for the specific tests, and retain the ESR example for
integration.