repo_synchroniser: push each reference separately (bug 1967265) (#57)
* retry: add usage examples to docstring
* repo_synchroniser: push each reference separately
* test_repo_synchroniser: test single-branch pushes
* test: use dedicated branch name to test branch mapping

Note on the last point:

The test configuration is used both for unit and integration tests. So
far, we have been using the example of an ESR to test dynamic
replacements in URL, branch and tags, which has been working as expected
across all tests.

However, with the change separating each ref into their own
push. The configuration resulting in pushes to both the `default` and
`mozilla-esr*` branches creates an issue, as the second push does not
contain any new commit, and gets rejected by Mercurial.

This is not a pattern we currently want to support, so the simplest
approach is to split the parameters so we have dedicated branch-mapping
config for the specific tests, and retain the ESR example for
integration.